fixed bug with wrong data type

+++ src/utils/API.js
@@ -28,8 +28,8 @@ export const uploadCarImage = async (img) => {
 
 export const createReport = async (lat, long, mediaURL, platesNumber, comment) => {
   const dataReport = {
-    lat: parseInt(lat, 10),
-    long: parseInt(long, 10),
+    lat: parseFloat(lat, 10),
+    long: parseFloat(long, 10),
     mediaURL: mediaURL,
     platesNumber: platesNumber,
     comment: comment,
